fix: grayout for nav and setting, refactor from cognito.ts to userAuth.ts
