+++ src/entities/User/ui/UserAvatarTooltip/UserAvatarTooltip.module.scss
@@ -27,6 +27,11 @@
 			}
 		}
 
+		&.disabled {
+			pointer-events: none;
+			color: var(--color-neutral);
+		}
+
 		.text {
 			display: block;
 			font: var(--font-s);
