+++ src/features/Nav/model/NavItems.ts
@@ -3,27 +3,32 @@ import { RoutePath } from 'app/providers/AppRouter';
 export type NavItemType = {
   path: string;
   text: string;
+  disabled?: boolean;
 };
 
 export const NavItemsList: NavItemType[] = [
-  // {
-  //   path: RoutePath.dashboard,
-  //   text: 'Dashboard',
-  // },
+  {
+    path: RoutePath.dashboard,
+    text: 'Dashboard',
+    disabled: true,
+  },
   {
     path: RoutePath.funnel,
     text: 'Funnel',
   },
-  // {
-  //   path: RoutePath.finances,
-  //   text: 'Finances',
-  // },
-  // {
-  //   path: RoutePath.shifts,
-  //   text: 'Shifts',
-  // },
-  // {
-  //   path: RoutePath.inventory,
-  //   text: 'Inventory',
-  // },
+  {
+    path: RoutePath.finances,
+    text: 'Finances',
+    disabled: true,
+  },
+  {
+    path: RoutePath.shifts,
+    text: 'Shifts',
+    disabled: true,
+  },
+  {
+    path: RoutePath.inventory,
+    text: 'Inventory',
+    disabled: true,
+  },
 ];
