+++ src/features/Nav/ui/NavItem/NavItem.module.scss
@@ -29,5 +29,9 @@
 				width: 100%;
 			}
 		}
+		&.disabled {
+			pointer-events: none;
+			color: var(--color-neutral);
+		}
 	}
 }
