+++ src/shared/ui/Progress/Progress.types.ts
